% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorhub-package.R
\docType{package}
\name{colorhub-package}
\alias{colorhub}
\alias{colorhub-package}
\title{colorhub: Convert Color Between Color Systems}
\description{
A user-friendly collection of functions for convert colors between color systems.
}
\keyword{internal}
