% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_trans.R
\name{MakeTrans}
\alias{MakeTrans}
\title{Produce color transparent}
\usage{
MakeTrans(color, percent = 0)
}
\arguments{
\item{color}{STRING for a color name of Hex code.}

\item{percent}{NUMERIC value between 0 and 1 for transparent.}
}
\value{
\strong{MakeTrans()} returns Hex code.
}
\description{
\strong{MakeTrans()} is a function for simply producing semi-transparent color
using color name.
}
\examples{
library(colorhub)
colorTrans <- MakeTrans("red", percent = 0.7)

colorTrans

}
\references{
Ying Wei (2021). Colors in R. retrieved from:
\emph{http://www.stat.columbia.edu/~tzheng/files/Rcolor.pdf}
}
\seealso{
\code{\link{SmrzTrans}}
}
\author{
Enoch Kang
}
