% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_trans.R
\name{SmrzTrans}
\alias{SmrzTrans}
\title{Summary of output of MakeTrans}
\usage{
SmrzTrans(data)
}
\arguments{
\item{data}{STRING for output of \strong{MakeTrans()}.}
}
\value{
\strong{SmrzTrans()} returns Hex code.
}
\description{
\strong{SmrzTrans()} is a function for summarizing output of \strong{MakeTrans()}.
}
\examples{
library(colorhub)
colorTrans <- MakeTrans("red", percent = 0.7)

colorTrans

SmrzTrans(colorTrans)

}
\seealso{
\code{\link{MakeTrans}}
}
