% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transparent.R
\name{Transparent}
\alias{Transparent}
\title{Produce color transparent}
\usage{
Transparent(color, percent = 0)
}
\arguments{
\item{color}{STRING for a color name of Hex code.}

\item{percent}{NUMERIC value between 0 and 1 for transparent.}
}
\value{
\strong{Transparent()} returns Hex code.
}
\description{
\strong{Transparent()} is a function for simply producing semi-transparent color
using color name.
}
\examples{
library(colorhub)
colorTrans <- Transparent("red", percent = 0.7)

colorTrans

}
\references{
Ying Wei (2021). Colors in R. retrieved from:
\emph{http://www.stat.columbia.edu/~tzheng/files/Rcolor.pdf}
}
\author{
Enoch Kang
}
